Fix paths due to updated qmake non-platform-specific paths.
Send the new TF to the GPU on data changes.

This fixes a tiny bug where loading a dataset which has an
associated .2dt file, and then selecting 2D TF rendering mode,
causes a blank window to appear.  It seemed as if the TF never got
pushed down to the GPU.