Adds variables and methods needed for clipping.
Just initializes them to default values in this commit.  The
methods are not called in this commit.

Rotating and translating a clip plane works very similarly to
rotating and translating the dataset.  We create a new arcball for
managing these transformations.  We also have some new matrices
for accumulating them together.  These apply to a PLANE<T> member
which is eventually passed into Tuvok.