Appropriately keep track of state.
Reset the rotation matrix when appropriate.
Move the tracking of the clip planes locking state to the rotation
and translation methods inside the event handlers; this keeps
everything in ImageVis3D, for one, but also keeps the logic close
to where it matters: if the volume rotates && the clip plane is
locked, the clip plane rotates too.  Thus one would expect to find
this logic in the `volume rotation' function, which is where it
now resides.