Ignore mouse/keyboard events during cleanup.
Due to a race condition in our single-threaded code (not a typo :),
we can receive a mouse movement event while our we are telling our
renderer to clean up after itself.  This disables Qt mouse/keyboard
events for a render window while it is dying, preventing our
handlers from ever being called in this case && thus allowing them
to assume that m_Renderer is valid.