Fix more build issues on Win32.
DX is probably broken, though it's likely been broken for a while.
Our GPU MemMgr has depended on GL code for a bit, so any code path
that uses it but not OpenGL is going to die.  This fixes the build
issue but it certainly won't work.

Fixed an issue with the help dialog too; missing an include.