Don't open a render window if we can't correctly render into it.
This adds some checks to make sure we respect the initialization
failure of the renderer.  It's a bit of a pain because when this
occurs, because we can still get other interaction events
(seemingly out of order) and we need to make sure those are
ignored if the window creation failed (at least until the `close'
event gets delivered.)

We should still continue if the GPU reports that the supported 3D
texture size is smaller than we require.  (Many drivers appear to
report that they don't support the texture size we need, but don't
give any error / work flawlessly if we just do it anyway.)  This
is only for the case where the GPU doesn't support something we
absolutely require, such as GLSL.

To test this, set m_bRenderSubsysOK to false near the end of
InitializeRenderer.