Allow closing the stereo tool widget when disabled.
We would disable the entire widget previously, which meant one
couldn't use the tiny `x' to close the window if no dataset was
loaded.  This changes the code to disable the individual entities
in the widget, but leaves the widget itself always enabled.  So,
one still can't play with stereo rendering parameters while there
is no active dataset, but the window is always closable (without
using the workspace menu).