Turn off UI elements first when a RenderWindow is closed.
Emits the signal to disable the render window before trying to
clean up the renderer.  We had a bit of a race between closing a
window and trying to use the menu to clone the previous window.
Opening the menu seems to force events to happen, particularly
paint events.  This means we could end up trying to paint while
our render window was dying.

Now we emit the signal first, and the first thing we do when
receiving that signal is to disable the Qt widget.  This causes
future events for the widget to be ignored.