Updated version number.
This is huge because it changes all of our version number parsing
and related code.  Our version numbers are now a series of numbers
instead of just a single floating point value.  This makes
comparing them a bit more complicated.  Further, we used to just
grep out the string version number and use in nightly scripts, but
that would include quotes now so it got a bit more involved.