Fix manual downloading/lookup.
Mac was pushd'd around before it went to look for the manual,
breaking the underlying grep of the Defines header.  Further, the
grep gave back a quoted string even when it did work, confusing
subsequent shell code.

We now make sure to lookup the manual before cd'ing around, strip
quotes off when we're looking up the manual, and finally quote the
manual's variable every time we use it.