Add release defines to get around debug iterator issues.
Too much of a pain to go through every file; we just unset _SECURE_SCL
and then reset it to off via command line parameters when we build in
release mode.  Thus it automatically applies to everything, and it won't
be a big deal to make sure the ordering is sound as we add files.

Also serialize the project's builds.  There's a race condition in the
mspdbsrv service or library, which is triggered when the debug iterator
mismatch compile errors occur.  If you have hit it, you must *manually*
remove your build directories and rebuild (a "clean rebuild" won't fix it).