Fix a crash when a dataset failed to convert.
When the dataset fails to convert, we'll get a `window is active'
event, but our window initialization will fail due to the missing
dataset.  Then we'll signal we want to die and thereafter get a
`window closing' event, which will end up calling this method.

Since RWA failed, we won't have a renderer && would previously
have deref'd a null pointer.