Move MainWindow show up && check for crash down.
This isn't ideal -- we'd like to check for a crash on the previous
run ASAP -- but moving it higher caused annoying flicker.
Essentially, this fixes the same problem we had with the welcome
screen: if the modal child becomes visible before its parent, it
ends up *behind* the parent under Qt/Cocoa.