Make progress update more responsive.
Linux doesn't seem to like to update the progress bar often
enough.  With a repaint + a processEvents, rendering actually
grinds to a halt with all of the events.  An update with a
processEvents doesn't get swamped with events (as here), though it
doesn't visibly update as often as we'd like.