Fix test code from debugging .deb creation script.
Hacked it to grab the debian dir from the local copy, so I
wouldn't have to make a commit every time I changed a single line
and wanted to test it.

Seems like I'll want that use case again someday, but in the
general case we should just use the (not)debian info that's
already in the repository.