Fix `Basic' mode 3-vertex "other" polygon rendering.
You'd think that we'd never have an "Other" polygon with fewer
than 5 vertices, as they would be subsumed by the pseudotri and
rectangle cases.

However a 3-vertex "Other" polygon can occur -- if we have a
triangle that isn't locked to the bottom, it's not a pseudotri and
thus gets dumped in the "other" bin.

This fixes a crash with a 2D TF I had lying around which had such
a real-not-pseudo triangle.