Ask the user if they want to proceed despite poor support.
We detect systems that don't support the required 3D texture size.
Unfortunately, some drivers report that they do not support the 3D
texture size we require... yet if you force them to use it, there
are no issues.  Other drivers actually do report an accurate
maximum 3D texture size, and will crash if you try to use
something larger.

Instead of trying to force it, leave the decision up to the user.
If we detect an unsupported configuration, ask the user if we
should continue while noting that doing so is dangerous.

I haven't actually tested the failure case here, but I'll find a
machine and test before release.