Allow LoadDataset to be interactive.
Practically, this just means the raw conversion dialog will pop up
if necessary.  It is a bit strange to get a dialog while running a
script, which you would normally expect to be non-interactive, but
without this "open" is essentially useless with raw data.  In the
face of doing the wrong thing or doing the surprising thing, this
opts for the latter.