Remove "upload" from the debug log UI.
The logs that are uploaded are completely useless, unless we have
told a specific user to upload a log and can correlate a
conversation with an uploaded log.  I haven't done that in a long
while, and we get a lot of junk debug logs uploaded.

If users want to report an issue, they should use the bug
reporting dialog.  That adds information, or if we're lucky, at
least contact info, that we can use to figure out what the real
issue is.