Add scripts for building modern Qt on windows.
The most recent release of Qt doesn't seem to be a huge
hassle in terms of getting the right options while compiling.
We can use most of the options from our other builds.

This also adds a script which will rm/extract the Qt tarball
and then run the actual build script.