Mac progress bar / window update fixes.
This fixes the long standing crash we've had when closing windows
on the Mac.  The issue was that Qt would try to deliver `please
render now' events *while* the window was dying.  I disabled the
window as the very first action we take on a close event,
preventing the delivery of any future draw events.  I also hid the
window, which makes it "feel" like ImageVis3D closed the window
much more quickly than it actually can.

Our progress update code has always been strange, because of not
updating enough frequently enough on Mac.  I think our best bet is
to hit a bunch of Qt objects in a method with an 'update' and then
force a processEvents at the end of the method.  That seems in line
with Qt's model for efficiency, and still makes sure things happen on
a Mac -- without delivery of an annoying bunch of messages.