Think of all wheel motion as a single `click'.
Regardless of speed, wheel movement will move us one slice when
using the 2D views.  Some mice do not obey the semi-standard
values for a number of degrees which have moved.  Fortunately, in
our slice views we only care about the direction of movement &&
always want to move one increment anyway, so here we just see
whether the movement is negative or positive and ignore any
other aspect of it.