Make script mode entirely synchronous.
For one, it's incredibly valuable to be able to test that our GUI
code does not rely on our timer firing at a particular time.  This
also makes the test suite render at high quality all of the time,
instead of only in the case that it has enough time to do so.