Revamp window init, removing dead code.
This gets rid of most of our hacky "force active" code, at long
last.  We don't need it anymore because we can guarantee
synchronous initialization && delivery of the events we need.

We do have a processEvents at the end here.  Some client code
assumes that when this method returns, the RenderWindow is open
and valid, i.e. that initialization is complete.  The
processEvents ensures that the initializeGL slot has fired.  We
might want to change it to ignore user actions..