Don't crash when the data file cannot be opened.
In particular, when the UVF version is bad.  If that happens,
we'll fail creation of the render window as well.

This really needs to be revamped, the error message is terrible
and confusing.  We probably want the IO subsystem to throw an
exception, because it's the only piece of code that knows what's
wrong, and accessing it from here manually would require a messy
accessor path.

Not right now though.