Fix capturing of 3D regions.
A previous Tuvok change made Tuvok bail if rendering took too long,
even in capture mode.  This is needed; we need to be able to
interrupt Tuvok, in the case that a user clicks a "cancel" button,
for example.  However, it means that any code which must ensure
Tuvok has finished rendering, like frame capturing, must spin in a
loop calling Tuvok and then checking if cancel was called or
doing any other updates.

This implements the above scheme for 3D rotations.  First, we
disable our global timer while we are capturing a rotation.  This
lets us handle events in a completely synchronous manner, giving
us more control.  Then we set capture mode and spin, giving Tuvok
a timeslice and then checking if the user hit cancel.  This
actually fixes a long standing bug; previously, IV3D could only
cancel between (Tuvok) frames when doing a capture; now we can
cancel between every time slice.  This is a big deal when
rendering the richtmyer at full res (648 bricks), for example.

We do another spin in CaptureSequenceFrame's implementation.  This
is because that code is called from scripting mode, and that mode
needs to be sure rendering is complete as well.  As far as 3D
rotation is concerned, this is a no-op: we've already finished a
complete render, so CheckForRedraw will end up false and we won't
end up spinning.  There is a small amount of code duplication, but
the alternative was letting RW know about PleaseWait, which is
dumb.  We can clean it up post-release.

As an added bonus, we now update the progress bar while rendering
rotations.  This is pretty nice, because otherwise we only get the
frame number in the animation, and a single frame can take minutes
when there are a significant number of bricks.

The need for these interactivity improvements was made more
critical because I forced capture mode to render the highest LOD
earlier.  Previously, it was taken for granted that the LOD
calculation mechanism would chose a fairly coarse (quick to
render) LOD.