Check for a dead renderer when applying settings.
We attempt to apply settings to all open render windows, if the pointers are
non-null.  I hit a case where the render window still existed but the renderer
was null, as can temporarily happen when we try to open a RW and then realize
that the system's OGL support is insufficient.

AFAIR, we should synchronously delete the renderer and close the RW, so I'm
not entirely sure how this case arose.  That said, this fixes the problem,
and it's probably good to be paranoid about these kinds of things anyway.