Update search to find manual on Linux.
We'll now actually search for the manual, and if we can't find it
we'll at least fail gracefully and let the user know what was
wrong.  We also try a list of PDF-reading programs, instead of
hardcoding a single one, so we should find at least one available
reader on the system.

The list is still hardcoded though; we should use shared-mime-info
to figure out which program the user wants to use in the future.