Change default directory for loading/saving 1D TFs.
Instead of using the last directory we saved a TFqn into, attempt
to use the directory of the currently-opened data set.  This is
mostly where one wants to store things anyway, because we autoload
transfer functions from this directory.

If for some reason the above doesn't work, we'll still attempt to
fall back to the previously-used directory.