added mutex to protect the paint method of the renderwindow from being called more than once at a time
Explanation of the issue:
How can this happen, aren't we (mainly) single threaded? The problem is our message reporting API sometimes calls Qt's processEvents. Thus if we have for instance a PleaseWait dialog box open and call our MESSAGE (or WARNING etc.) methods somewhere in the paint routine, we return control back to Qt, that might invoke another paint call, once that call would be finished our first call would resume painting. As this may lead to all kinds of issues, we now explicitly catch multiple paint calls with a mutex and return emediatly if a paint is already in proress.