Fix visibility warning on Mac.
The Qt binaries that ship w/ RHEL appear to be broken in that they
screw up symbol visibility on some of the Qt symbols.  To get a
no-warning build on Linux, then, I inserted a hack to remap gcc's
visibility in just the specific places we were having issues.

That fix was incorrect; it should be a Linux-only hack (preferably
a RH-only hack, to be honest) but I incorrectly applied it to all
platforms.  It got changed to be a hack for non-windows later,
which was better but still left the Mac giving us warnings about
symbol visibility at link time.

This commit switches the hack to only affect compilation when using
both Linux and GNU compilers.