Do no try to build IO tests on Darwin.
The compiler there is too old; it's very buggy w.r.t. some of the
later TR1 stuff.  Particularly the random number generation code,
which is what bites us here.

For now, just give up on Darwin.  Maybe in a few eternities when
LLVM's C++ support is ready, we can turn it on.