Block the global redraw timer while asking to switch renderers.
Otherwise we can get a 'CheckForRedraw' event while the user is
selecting whether or not they want to switch their renderer type.
If this happens and then the user says 'yes', then we can end up
drawing while we are killing and reinitializing a renderer, which
leads to all sorts of crashes.