Add a relative isovalue setting, for locked windows.
If we have an 8bit dataset locked to a 12bit dataset ranging from
0 to 4096, when we set the 8bit DS' isovalue to 128, we want to
set the 12bit DS' isovalue to 2048, not 128.

This adds a relative set function, which calls the (new)
similarly-named AbstrRenderer function. The catch is that the
values are supposed to be in the range [0,1], instead of in the
range of the data set. This allows us to interpret it as a
percentage and in turn re-interpret it appropriately from within
tuvok.