Fixed Qt build issue.
Removed the experimental opengl graphics system. See
http://doc.qt.nokia.com/4.7-snapshot/configure-options.html. Note that
it says both OpenGL and OpenVG are experimental graphics systems. After
reading up on it, it appears that the graphicssystem renders Qt
widgets -- so OpenGL or OpenVG would bypass native widgets.

See:
http://labs.qt.nokia.com/2010/01/06/qt-graphics-and-performance-opengl/.

I tracked the library issues on Mac down to the
libqglgraphicssystem*.dylib libraries. With DYLD_PRINT_LIBRARIES=1, I
compared the library output from a correct executable against that of a
broken executable. The only notable difference was that the correct
executable *failed* to find libqglgraphicssystem.