Add c++11 flag to .pro files.
Instead of relying on the user providing it when running qmake.
Couldn't figure out the ideal way of doing this, which would be
something like:

  if(the compiler is gcc) {
    add the flag "-std=c++0x"
  }

Turns out clang would also accept such a flag.  And the intel
compiler would too.  So that 'if' is a little too narrow, but the
point is: we should be testing what the *compiler* is, not the
*platform*, when adding this flag.

Couldn't figure out how to do that with qmake ... this will
suffice for now, other fixes welcome.