Fixed locked clip plane rotation.
Matrix multiplication was out of order. Suppose we had 3 rotations given
to us in the order: R1, R2, and R3. Instead of performing the correct
R1 * R2 * R3, we multiplied R3 * R2 * R1.

This breaks non-locked clipping plane rotation because the toClipping plane
variable is being used incorrectly. toClipping plane is really
"fromClippingPlane (to world)" when the clipping plane is not locked to
an object (it is fine when locked to an object however).
Will be fixed in the next commit.