assert our tfqn buffer is as wide as our histogram.
I hit an unfortunate case where we were attempting to redraw the
tfqn *while* doing another operation.  The other operation
apparently modified the transfer function, and then tried to
print a debug message... which caused us to process events,
including redrawing the tfqn, which accessed the modified transfer
function as if it was still the same size.

Can't reproduce it of course.  Hopefully someone else will catch
it via this assert.