Resolving exit crash in llvm_gcda_start_file.
Clang attempts to open gcda coverage files and crashes.

This is somewhat related to http://llvm.org/bugs/show_bug.cgi?id=12358r;
but I verified that relative paths are being used to open coverage
files in IV3D. The problem persists in that llvm_gcda_start_file is not
recovering gracefully when unable to open a coverage file.

Currently researching whether there is a special way to handle code
coverage for apps in /Applications.