Fix error report upload dialog.
It wasn't properly notifying users when it completed.  I guess the
commandFinished signal doesn't properly get emitted anymore, or at
least not with the appropriate ID.

Instead register a new function for the signal 'done', which gets
called at the end of the entire series of FTP instructions no
matter what--fail or not.  Then control the 'success!' or
'failure!' messages from that slot's implementation.