Add OpenMP to build flags in debug mode.
Hack to work around an issue with gcc warning on unknown pragmas
(sigh).  There might be a warning option we can fiddle with so
that we can leave on Werror without turning on OMP, will need to
look into that.

Anyway this should fix the build for now.

I also cleaned up the script a bit:
  . QLF was never set to anything.
  . QMAKE_LFLAGS used COVERAGE instead of LDFLAGS directly.
  . 32 bit flag added in a strange place/way.
  . Maybe something else...