Add a scripting command to change the transfer function.
This one was a bit messy because I wanted the function to be
attached to a render window (at least for now).  But RenderWindows
and even Renderers don't actually have any notion of loading
transfer functions; instead we just load them elsewhere and say,
"hey, btw, here's a buffer.  It's a transfer function." and then
the underlying data is just set in the renderer.

So this makes a call to the mainwindow which in turn calls right
back to the RWindow... we should probably revisit this.