Auto-choose the renderer in basic mode.
If the user is not using advanced settings, choose which renderer
they should use based on a (very simple) heuristic.

We really need to rework how renderers get initialized sometime,
this is a bit hacky.  We need to wait until we init OpenGL so we
can query some things, which means we need to create a bogus
renderer and then re-set it.  Ugh.