Add RenderWindow lua call for fiddling with the clipping plane.
I think this has bad semantics right now... namely that it just
sets it raw and completely ignores what the dataset's rotation is,
but it should really only track that if it's locked anyway, but it
really doesn't matter much because we need to simplify our RW's
Lua interface anyway. Probably by exposing more AbstrRenderer
calls.