Fix up windows batch scripts.
Remove the confusing mass of Qt-building scripts.  Just have/keep
one.  Qt's "configure" also seems to now actually work correctly on
windows, so we can disable a lot more stuff.

Change up the build scripts to "Build" instead of "Rebuild".  For
CI, they're always building in a fresh checkout, so it's the same
anyway.  And now the scripts are usable if you're just doing dev
locally.